+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { createUseFetch } from './use-fetch';
 export { createUseMutation } from './use-mutation';
 export { createUsePaginate } from './use-paginate';
+export { useIsFetching, useIsMutating } from './use-is-fetching';